Major refactor and remove some options
 - Save config settings (non-defaults)
 - Remove the CoffeeScript option
 - Remove gruntfile adding
 - Accepts csv or array

Fixes #59

Command to run is added to the package.json file. In keeping with moving
away from specific build systems, the Gruntfile additions are being
removed. There are enough tutorials for people to be able to add this to
their build system.

Remove CoffeeScript as it can be added in another generator that
composes using this one